NIFI-10212 small fixes
