+++ nifi-assembly/pom.xml
@@ -666,13 +666,13 @@ language governing permissions and limitations under the License. -->
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smb-client-provider-api-nar</artifactId>
+            <artifactId>nifi-smb-client-api-nar</artifactId>
             <version>1.17.0-SNAPSHOT</version>
             <type>nar</type>
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smb-client-provider-nar</artifactId>
+            <artifactId>nifi-smbj-client-nar</artifactId>
             <version>1.17.0-SNAPSHOT</version>
             <type>nar</type>
         </dependency>
