--- nifi-nar-bundles/nifi-smb-bundle/nifi-smb-client-provider-api/src/main/java/org/apache/nifi/services/smb/SmbClientService.java
@@ -1,18 +1,0 @@
-package org.apache.nifi.services.smb;
-
-import java.io.OutputStream;
-import java.util.stream.Stream;
-
-/**
- *  Service abstraction for Server Message Block protocol operations.
- */
-public interface SmbClientService {
-
-    Stream<SmbListableEntity> listRemoteFiles(String path);
-
-    void createDirectory(String path);
-
-    OutputStream getOutputStreamForFile(String path);
-
-    void close();
-}
