+++ nifi-nar-bundles/nifi-smb-bundle/nifi-smb-nar/pom.xml
@@ -37,12 +37,12 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smbj-client-provider</artifactId>
+            <artifactId>nifi-smbj-client</artifactId>
             <version>1.17.0-SNAPSHOT</version>
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smb-client-provider-api-nar</artifactId>
+            <artifactId>nifi-smb-client-api-nar</artifactId>
             <version>1.17.0-SNAPSHOT</version>
             <type>nar</type>
         </dependency>
