+++ nifi-nar-bundles/nifi-smb-bundle/nifi-smb-processors/pom.xml
@@ -28,7 +28,7 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smb-client-provider-api</artifactId>
+            <artifactId>nifi-smb-client-api</artifactId>
             <version>1.17.0-SNAPSHOT</version>
             <scope>provided</scope>
         </dependency>
@@ -75,7 +75,7 @@
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smbj-client-provider</artifactId>
+            <artifactId>nifi-smbj-client</artifactId>
             <version>1.17.0-SNAPSHOT</version>
             <scope>test</scope>
         </dependency>
