+++ nifi-nar-bundles/nifi-smb-bundle/nifi-smbj-client-nar/pom.xml
@@ -22,7 +22,7 @@
         <version>1.17.0-SNAPSHOT</version>
     </parent>
 
-    <artifactId>nifi-smbj-client-provider-nar</artifactId>
+    <artifactId>nifi-smbj-client-nar</artifactId>
     <packaging>nar</packaging>
     <properties>
         <maven.javadoc.skip>true</maven.javadoc.skip>
@@ -32,13 +32,13 @@
     <dependencies>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smb-client-provider-api-nar</artifactId>
+            <artifactId>nifi-smb-client-api-nar</artifactId>
             <version>1.17.0-SNAPSHOT</version>
             <type>nar</type>
         </dependency>
         <dependency>
             <groupId>org.apache.nifi</groupId>
-            <artifactId>nifi-smbj-client-provider</artifactId>
+            <artifactId>nifi-smbj-client</artifactId>
             <version>1.17.0-SNAPSHOT</version>
         </dependency>
     </dependencies>
