+++ nifi-nar-bundles/nifi-smb-bundle/pom.xml
@@ -26,10 +26,10 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>nifi-smb-client-provider-api</module>
-        <module>nifi-smb-client-provider-api-nar</module>
-        <module>nifi-smbj-client-provider</module>
-        <module>nifi-smbj-client-provider-nar</module>
+        <module>nifi-smb-client-api</module>
+        <module>nifi-smb-client-api-nar</module>
+        <module>nifi-smbj-client</module>
+        <module>nifi-smbj-client-nar</module>
         <module>nifi-smb-processors</module>
         <module>nifi-smb-nar</module>
     </modules>
